refactor: move folding functions to buffers module
